+++ collection.go
@@ -22,9 +22,11 @@ package leia
 import (
 	"context"
 	"crypto/sha1"
+	"encoding/json"
 	"errors"
 	"fmt"
 
+	"github.com/piprate/json-gold/ld"
 	"github.com/tidwall/gjson"
 	"go.etcd.io/bbolt"
 )
@@ -45,8 +47,6 @@ func documentCollectionByteRef() []byte {
 
 // Collection defines a logical jsonCollection of documents and indices within a store.
 type Collection interface {
-	// Name returns the name of the collection
-	Name() string
 	// AddIndex to this jsonCollection. It doesn't matter if the index already exists.
 	// If you want to override an index (by path) drop it first.
 	AddIndex(index ...Index) error
@@ -73,11 +73,8 @@ type Collection interface {
 	// IndexIterate is used for iterating over indexed values. The query keys must match exactly with all the FieldIndexer.Name() of an index
 	// returns ErrNoIndex when no suitable index can be found
 	IndexIterate(query Query, fn ReferenceScanFn) error
-	// ValuesAtPath returns a slice with the values found at the given query path
+	// ValuesAtPath returns a slice with the values found by the configured valueCollector
 	ValuesAtPath(document Document, queryPath QueryPath) ([]Scalar, error)
-
-	// todo
-	DB() *bbolt.DB
 }
 
 // ReferenceFunc is the func type used for creating references.
@@ -96,10 +93,12 @@ func defaultReferenceCreator(doc Document) Reference {
 }
 
 type collection struct {
-	Name      string `json:"name"`
-	db        *bbolt.DB
-	IndexList []Index `json:"indices"`
-	refMake   ReferenceFunc
+	name              string
+	db                *bbolt.DB
+	indexList         []Index
+	refMake           ReferenceFunc
+	documentProcessor *ld.JsonLdProcessor
+	valueCollector    valueCollector
 }
 
 func (c *collection) NewIndex(name string, parts ...FieldIndexer) Index {
@@ -371,21 +370,110 @@ func (c *collection) Get(key Reference) (Document, error) {
 }
 
 func (c *collection) documentBucket(tx *bbolt.Tx) *bbolt.Bucket {
-	bucket := tx.Bucket([]byte(c.Name))
+	bucket := tx.Bucket([]byte(c.name))
 	if bucket == nil {
 		return nil
 	}
 	return bucket.Bucket(documentCollectionByteRef())
 }
 
-// ValuesAtPath returns a slice with the values found at the given JSON path query
-func (c *collection) ValuesAtPath(document Document, jsonPath string) ([]Scalar, error) {
+// valueCollector is responsible for going through the document and finding the Scalars that match the Query
+type valueCollector func(collection *collection, document Document, queryPath QueryPath) ([]Scalar, error)
+
+func JSONPathValueCollector(_ *collection, document Document, queryPath QueryPath) ([]Scalar, error) {
+	jsonPath, ok := queryPath.(jsonPath)
+	if !ok {
+		return nil, ErrInvalidQuery
+	}
+
 	if !gjson.ValidBytes(document) {
 		return nil, ErrInvalidJSON
 	}
-	result := gjson.GetBytes(document, jsonPath)
+	result := gjson.GetBytes(document, string(jsonPath))
 
 	return valuesFromResult(result)
+}
+
+func JSONLDValueCollector(collection *collection, document Document, queryPath QueryPath) ([]Scalar, error) {
+	termPath, ok := queryPath.(termPath)
+	if !ok {
+		return nil, ErrInvalidQuery
+	}
+
+	if len(termPath.terms) == 0 {
+		return []Scalar{}, nil
+	}
+
+	var input interface{}
+	if err := json.Unmarshal(document, &input); err != nil {
+		return nil, err
+	}
+
+	expanded, err := collection.documentProcessor.Expand(input, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return valuesFromSliceAtPath(expanded, termPath), nil
+}
+
+func valuesFromSliceAtPath(expanded []interface{}, termPath termPath) []Scalar {
+	result := make([]Scalar, 0)
+
+	for _, sub := range expanded {
+		switch typedSub := sub.(type) {
+		case []interface{}:
+			result = append(result, valuesFromSliceAtPath(typedSub, termPath)...)
+		case map[string]interface{}:
+			result = append(result, valuesFromMapAtPath(typedSub, termPath)...)
+		}
+	}
+
+	return result
+}
+
+func valuesFromMapAtPath(expanded map[string]interface{}, termPath termPath) []Scalar {
+	// JSON-LD in expanded form either has @value, @id, @list or @set
+	if termPath.IsEmpty() {
+		if value, ok := expanded["@value"]; ok {
+			return []Scalar{MustParseScalar(value)}
+		}
+		if id, ok := expanded["@id"]; ok {
+			return []Scalar{MustParseScalar(id)}
+		}
+		if list, ok := expanded["@list"]; ok {
+			castList := list.([]interface{})
+			scalars := make([]Scalar, len(castList))
+			for i, s := range castList {
+				scalars[i] = MustParseScalar(s)
+			}
+			return scalars
+		}
+		if set, ok := expanded["@set"]; ok {
+			castSet := set.([]interface{})
+			scalars := make([]Scalar, len(castSet))
+			for i, s := range castSet {
+				scalars[i] = MustParseScalar(s)
+			}
+			return scalars
+		}
+	}
+
+	if value, ok := expanded[termPath.Head()]; ok {
+		// the value should now be a slice
+		next, ok := value.([]interface{})
+		if !ok {
+			return nil
+		}
+		return valuesFromSliceAtPath(next, termPath.Tail())
+	}
+
+	return nil
+}
+
+// ValuesAtPath returns a slice with the values found at the given JSON path query
+func (c *collection) ValuesAtPath(document Document, queryPath QueryPath) ([]Scalar, error) {
+	return c.valueCollector(c, document, queryPath)
 }
 
 func valuesFromResult(result gjson.Result) ([]Scalar, error) {
