+++ plan_test.go
@@ -32,7 +32,7 @@ func TestFullTableScanQueryPlan_execute(t *testing.T) {
 		_, c := testCollection(t)
 		queryPlan := fullTableScanQueryPlan{
 			queryPlanBase: queryPlanBase{
-				collection: &c,
+				collection: c,
 			},
 		}
 
@@ -49,7 +49,7 @@ func TestFullTableScanQueryPlan_execute(t *testing.T) {
 		_ = c.Add([]Document{exampleDoc})
 		queryPlan := fullTableScanQueryPlan{
 			queryPlanBase: queryPlanBase{
-				collection: &c,
+				collection: c,
 			},
 		}
 
@@ -84,7 +84,7 @@ func TestIndexScanQueryPlan_Execute(t *testing.T) {
 		_, c, i := testIndex(t)
 		queryPlan := indexScanQueryPlan{
 			queryPlanBase: queryPlanBase{
-				collection: &c,
+				collection: c,
 				query:      New(Eq(NewJSONPath("path.part"), valueAsScalar)),
 			},
 			index: i,
@@ -104,7 +104,7 @@ func TestResultScanQueryPlan_Execute(t *testing.T) {
 		_, c, i := testIndex(t)
 		queryPlan := resultScanQueryPlan{
 			queryPlanBase: queryPlanBase{
-				collection: &c,
+				collection: c,
 				query:      New(Eq(NewJSONPath("key"), valueAsScalar)),
 			},
 			index: i,
@@ -158,7 +158,7 @@ func TestResultScanner(t *testing.T) {
 		err := db.View(func(tx *bbolt.Tx) error {
 			scanner := resultScanner([]QueryPart{Eq(NewJSONPath("main.nesting"), valueAsScalar)}, func(_ Reference, _ []byte) error {
 				return errors.New("failed")
-			}, &c)
+			}, c)
 
 			return scanner(nil, bytes)
 		})
